SpringTutorial/src/main/java/startspring/hello_spring/discount/RateDiscountPolicy.java
package startspring.hello_spring.discount;

import org.springframework.beans.factory.annotation.Qualifier;
import org.springframework.context.annotation.Primary;
import org.springframework.stereotype.Component;
import startspring.hello_spring.annotation.MainDiscountPolicy;
import startspring.hello_spring.member.Grade;
import startspring.hello_spring.member.Member;

@Component
@MainDiscountPolicy
public class RateDiscountPolicy implements DiscountPolicy{

    private  int discountPercent = 10;

    @Override
    public int discount(Member member, int price) {
        if (member.getGrade() == Grade.VIP) {
            return price * discountPercent / 100;
        } else {
            return 0;
        }
    }
}
